requirements/requirements.txt
Django==1.11.15
Pillow==5.3.0
djangorestframework==3.9.0
markdown==3.0.1
django-filter==2.0.0
# coreapi pyyaml django-rest-swagger 文档依赖
coreapi==2.3.3
pyyaml==3.13
django-rest-swagger==2.2.0
drf-yasg[validation]==1.11.0
django-cors-headers==2.4.0
apscheduler==3.5.3
django-apscheduler==0.2.13
tzlocal==1.5.1
djangorestframework-jwt==1.11.0
mysqlclient==1.3.13